Separate remote / local test (seems like running in parallel on travis)
